pyproject.toml
[project]
name = "n8n-mcp-server"
version = "0.1.0"
description = "MCP server for controlling self-hosted n8n instances"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
n8n-mcp = "n8n_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
